discord/requirements.txt
aiodns==2.0.0
aiofiles==0.6.0
aiohttp==3.7.4.post0
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
Brotli==1.0.7
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cycler==0.10.0
dash==1.13.3
dash-core-components==1.10.1
dash-html-components==1.0.3
dash-renderer==1.5.0
dash-table==4.8.1
decorator==4.4.2
defusedxml==0.6.0
discord.py==1.7.1
Django==3.2
entrypoints==0.3
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
idna==2.10
image==1.5.33
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jaraco.context==4.0.0
jedi==0.17.2
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
more-itertools==8.7.0
multidict==5.1.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.19.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.3
parso==0.7.1
pickleshare==0.7.5
Pillow==8.2.0
plotly==4.8.2
prometheus-client==0.9.0
prompt-toolkit==3.0.8
pycares==3.1.1
pycparser==2.20
Pygments==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
pywin32==300
pywinpty==0.5.7
pyzmq==20.0.0
requests==2.25.0
retrying==1.3.3
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.4.1
terminado==0.9.1
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wolframalpha==5.0.0
xmltodict==0.12.0
yarl==1.6.3
youtube-dl==2020.6.16.1
ypstruct==0.0.2
zipp==3.4.0
